tempo/requirements.txt
Flask
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
